Hover bug fixed
Occasionally a crumb would get stuck as being hovered event after the
mouse cursor has left the window. This is because Crumbs depended on
motion notify events to un-hover a crumb. If the cursor was flicked
across the crumb, it would get the notification and set the proper crumb
to hover, but when the mouse cursor left it would get stuck there,
because there wasn't a notification that the cursor left.

Now we use leave notification to ensure when the cursor is out of the
widget there is no hovered crumb.