Crumbs in now a subclass of Gtk.HBox
When children are added to it, it wraps those children in a
Gtk.ToggleButton. Crumbs will also draw a button-styled box around
itself.

Exactly the look I'm going for (with most Gtk+ themes). It is now
accessible using the keyboard. We get special styling for selected,
pressed, prelit, etc from crumbs automatically.

Scroll selection needs to be reimplemented, though.

Got rid of Crumb, because it's no longer necessary.